tests/requirements.txt
yandex-taxi-testsuite[postgresql-binary] >= 0.1.7.2
Faker >= 33.0.0
pydantic >= 2.10.0
